# Routing algorithm performance improvement