@mossblaser 2016-06-07
> Realistically, how big might this become? Is it worth investigating a more cache-like structure?

In practice: 1 entry (with a list of 1261 2-element tuples).

New entries are only ever added when the NER router is called with a new radius value. In practice the only reason someone would change this (or call route more than once anyway for that matter) is for the purposes of doing routing experiments. In such cases I've been doing order tens of runs and so the cache size should still be very small in practice.

I considered making things get evicted but basically I couldn't be bothered to implement it due to the above. What do you think?