@mossblaser 2016-06-07
OK so I think this is about as far with this as I'm happy to go at the moment... Headline outcomes of this PR:

* Routing is now 2-3x faster
* The returned `RoutingTree`s now take up half as much memory as before.

The most controversial part of this patch set is that swapping `set`s for `list`s in the `RoutingTree` structure is strictly a breaking change. In practice, the only code which creates `RoutingTree`s outside the router is probably in test code for the routing paper. I *think* that thanks to the wonders of duck typing these tests probably will continue to work though they ideally would be updated...