@mundya 2016-06-07
This LGTM, apart from wanting an extra newline.

I *think* I'm happy to let this breaking change pass as (a) I agree that duck-typing should probably, generally, mean it's not an issue, (b) it's a mostly internal change.